Extract appearance from WordPressAppDelegate
